requirement.txt
fastapi
uvicorn
requests
streamlit
langchain
langchain-community
langchain-openai
langchain-huggingface
langchain-chroma
langchain-classic
chromadb
arxiv
pypdf
tiktoken
sentence-transformers
openai
pydantic
python-dotenv